LayoutTests/http/tests/download/sandboxed-iframe-download-not-allowed-expected.txt

CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing
This test passes if no download is started.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
